quentin/requirements.txt
absl-py==0.11.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
traitlets==4.3.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
